Refactor Curator to use overmind
